requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-taggit==4.0.0
django-extensions==3.2.3
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
